move components around, make image display correctly based on theme
