+++ website/src/components/CreateProjectAnimation.js
